+++ src/Application/Import/ImportJobPage.php
@@ -1,0 +1,3 @@
+<?php
+namespace EventFlow\Application\Import;
+final readonly class ImportJobPage{/** @param list<ImportJobRecord> $jobs */public function __construct(public array $jobs,public ?int $nextAfterJobId){}}